Add murder.js to /charts
