--- src/charts/section01.js
@@ -1,614 +1,0 @@
-const hc003_chart = {
-    data: {
-        table: 'hc003_table'
-    },
-    chart: {
-        type: 'column'
-    },
-    title: {
-        text: 'Data extracted from a HTML table in the page'
-    },
-    yAxis: {
-        allowDecimals: false,
-        title: {
-            text: 'Units'
-        }
-    },
-    tooltip: {
-        formatter: function() {
-            return '<b>' + this.series.name + '</b><br/>' +
-                this.point.y + ' ' + this.point.name.toLowerCase();
-        }
-    }
-}
-
-const hc003_table = `
-    <table id="hc003_table" class="ordered striped">
-    <thead>
-        <tr>
-            <th></th>
-            <th>Jane</th>
-            <th>John</th>
-        </tr>
-    </thead>
-    <tbody>
-     <tr>
-            <th>Apples</th>
-            <td>3000</td>
-            <td>4000</td>
-        </tr>
-        <tr>
-            <th>Pears</th>
-            <td>2000</td>
-            <td>0</td>
-        </tr>
-        <tr>
-            <th>Plums</th>
-            <td>5000</td>
-            <td>11000</td>
-        </tr>
-        <tr>
-            <th>Bananas</th>
-            <td>1000</td>
-            <td>1000</td>
-        </tr>
-        <tr>
-            <th>Oranges</th>
-            <td>2000</td>
-            <td>4000</td>
-        </tr>
-    </tbody>
-</table>`
-
-
-
-
-
-const hc004_chart = {
-    data: {
-        table: 'hc004_table'
-    },
-    chart: {
-        type: 'column'
-    },
-    title: {
-        text: 'Data extracted from a HTML table in the page'
-    },
-    yAxis: {
-        allowDecimals: false,
-        title: {
-            text: 'Units'
-        }
-    },
-    tooltip: {
-        formatter: function() {
-            return '<b>' + this.series.name + '</b><br/>' +
-                this.point.y + ' ' + this.point.name.toLowerCase();
-        }
-    }
-}
-
-
-const hc004_table = `
-    <table id="hc004_table" class="ordered striped">
-    <thead>
-        <tr>
-            <th></th>
-            <th>Jane</th>
-            <th>John</th>
-        </tr>
-    </thead>
-    <tbody>
-     <tr>
-            <th>Apples</th>
-            <td>3000</td>
-            <td>4000</td>
-        </tr>
-        <tr>
-            <th>Pears</th>
-            <td>2000</td>
-            <td>0</td>
-        </tr>
-        <tr>
-            <th>Plums</th>
-            <td>5000</td>
-            <td>11000</td>
-        </tr>
-        <tr>
-            <th>Bananas</th>
-            <td>1000</td>
-            <td>1000</td>
-        </tr>
-        <tr>
-            <th>Oranges</th>
-            <td>2000</td>
-            <td>4000</td>
-        </tr>
-    </tbody>
-</table>`
-
-
-const hc005_chart = {
-    title: {
-        text: 'Monthly Average Temperature',
-        x: -20 //center
-    },
-    subtitle: {
-        text: 'Source: WorldClimate.com',
-        x: -20
-    },
-    xAxis: {
-        categories: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun',
-            'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'
-        ]
-    },
-    yAxis: {
-        title: {
-            text: 'Temperature (°C)'
-        },
-        plotLines: [{
-            value: 0,
-            width: 1,
-            color: '#808080'
-        }]
-    },
-    tooltip: {
-        valueSuffix: '°C'
-    },
-    legend: {
-        layout: 'vertical',
-        align: 'right',
-        verticalAlign: 'middle',
-        borderWidth: 0
-    },
-    series: [{
-        name: 'Tokyo',
-        data: [2000, 2323, 2323, 4000]
-    }, {
-        name: 'New York',
-        data: [7000, 8000, 9234, 2343]
-    }, {
-        name: 'Berlin',
-        data: [1000, 999, 1345, 1245]
-    }, {
-        name: 'London',
-        data: [3000, 4534, 2342, 2323, 3244]
-    }]
-}
-
-const hc_M_def_chart_150 =  {
-        chart: {
-            type: 'column'
-        },
-        title: {
-            text: ''
-        },
-        subtitle: {
-            text: ''
-        },
-        xAxis: {
-
-  categories:["Cook","Marion","Marshall","Peoria","St. Clair","Will","Winnebago"]
-,
-            title: {
-                text: null
-            }
-        },
-        yAxis: [{
-            title: {
-                text: 'Murders',
-            },
-            labels: {
-                overflow: 'justify'
-            },
-        },
-        {
-
-            title: {
-                text: 'Rate per 100,000',
-            },
-            labels: {
-                overflow: 'justify'
-            },
-            opposite: true
-
-
-        }
-        ],
-        tooltip: {
-            shared: true
-
-        },
-        plotOptions: {
-            column: {
-                dataLabels: {
-                    enabled: true
-                }
-            }
-        },
-				legend:{
-        enabled: false
-        },
-        credits: {
-            enabled: false
-        },
-        series: [{
-            name: 'Murders',
-            data: [534,4,16,1,32,18,22]
-
-        },
-        {
-            yAxis: 1,
-            name: 'Murder Rate per 100,000',
-            data: [10.18,10.37,8.54,8.32,12.04,2.63,7.62]
-        }
-        ]
-    }
-
-
-    const hc_M_def_chart_300 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'pie'
-        },
-        title: {
-            text: ''
-        },
-        tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}%</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color:  'black'
-                    }
-                }
-            }
-        },
-        series: [{
-            name: 'Offenders',
-            colorByPoint: true,
-            data: [{name:"Male",y:89.7},
-{name:"Female",y:9.2},
-{name:"Unknown",y:1.1}]
-        }]
-    }
-
-
-const hc_M_def_chart_500 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'pie'
-        },
-        title: {
-            text: ''
-        },
-   			subtitle: {
-            text: ''},        tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}%</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color: 'black'
-                    }
-                }
-            }
-        },
-        series: [{
-            name: 'Offenders',
-            colorByPoint: true,
-            data: [{name:"Black",y:74.4},
-{name:"White",y:24.6},
-{name:"Unknown",y:1.0}]
-        }]
-    }
-
-
-    const hc_M_def_chart_400 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'pie'
-        },
-        title: {
-            text: ''
-        },
-   			subtitle: {
-            text: ''},        tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}%</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color: 'black'
-                    }
-                }
-            }
-        },
-        series: [{
-            name: 'Victims',
-            colorByPoint: true,
-            data: [{name:"Male",y:88.5},
-{name:"Female",y:10.6},
-{name:"Unknown",y:0.9}]
-        }]
-    }
-
-
-const hc_M_def_chart_600 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'pie'
-        },
-        title: {
-            text: ''
-        },
-   			subtitle: {
-            text: ''},        tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}%</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color: 'black'
-                    }
-                }
-            }
-        },
-        series: [{
-            name: 'Victims',
-            colorByPoint: true,
-            data: [{name:"Black",y:76.0},
-{name:"White",y:21.7},
-{name:"Unknown",y:1.4}]
-        }]
-    }
-
-
-    const hc_M_def_chart_700 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'pie'
-        },
-        title: {
-            text: ''
-        },
-   			subtitle: {
-            text: ''},
-            tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}%</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color: 'black'
-                    }
-                }
-            }
-        },
-        series: [{
-            name: 'Arrestees',
-            colorByPoint: true,
-            data: [{name:"Male",y:94.5},
-{name:"Female",y:5.5}]
-        }]
-    }
-
-
-const hc_M_def_chart_800 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'pie'
-        },
-        title: {
-            text: ''
-        },
-   			subtitle: {
-            text: ''},
-            tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}%</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color: 'black'
-                    }
-                }
-            }
-        },
-        series: [{
-            name: 'Arrestees',
-            colorByPoint: true,
-            data: [{name:"Black",y:73.2},
-{name:"White",y:15.5},{name:"Hispanic",y:9.8},{name:"Other/Unknown",y:1.5}]
-        }]
-    }
-
-const hc_M_def_chart_900 = {
-        chart: {
-            plotBackgroundColor: null,
-            plotBorderWidth: null,
-            plotShadow: false,
-            type: 'line'
-        },
-        title: {
-            text: ''
-        },
-   			subtitle: {
-            text: ''},
-            tooltip: {
-            pointFormat: '{series.name}: <b>{point.y:.1f}</b>'
-        },
-        plotOptions: {
-            pie: {
-                allowPointSelect: true,
-                cursor: 'pointer',
-                dataLabels: {
-                    enabled: true,
-                    format: '<b>{point.name}</b>: {point.y:.1f} %',
-                    style: {
-                        color: 'black'
-                    }
-                }
-            }
-        },
-        xAxis: {
-        categories:
- [14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,29.74,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,54,55,56,57,58,59,60,61,62,63,64,65,66,68,72,73,77],
-            plotLines: [{
-                color: '#7A4695',
-                width: 2,
-                value: 11,
-                label: {
-                text: 'Median<br> Age = 25',
-                textAlign: 'right',
-                x: -5,
-                y: 240,
-                rotation: 0
-                },
-            }, {
-                color: '#7A4695',
-                width: 2,
-                value: 15.74,
-                label: {
-                text: 'Average<br> Age = 29.4',
-                verticalAlign: 'top',
-                textAlign: 'left',
-                rotation: 0,
-                y: 150
-                }
-            }],
-            title: {
-                text: 'Age of Arrestee'
-            }
-   },
-   yAxis:{
-   title: {
-   text: "Number of Arrests"
-   }
-   },
-				legend:{
-        enabled: false
-        },
-        series: [{
-            name: 'Number of Arrests',
-            data: [3,4,18,19,32,22,42,40,29,30,22,17,22,18,12,6,5,5,13,4,11,8,10,18,9,8,10,6,4,5,11,6,6,3,4,8,8,4,6,3,2,2,3,1,7,3,3,2,2,2,2,1,1,1,1,1,1]
-
-        }]
-    }
-
-const hc_M_def_chart_1100 = {
-        chart: {
-            type: 'column'
-        },
-        title: {
-            text: ''
-        },
-        subtitle: {
-            text: ''
-        },
-        xAxis: {
-
-  categories:["0 priors","1 prior","2-5 priors","6-10 priors","11-15 priors","16-20 priors","21+ priors"]
-,
-            title: {
-                text: null
-            }
-        },
-
-        yAxis: {
-            title: {
-                text: '',
-            },
-            labels: {
-                overflow: 'justify',
-                formatter: function () {
-                    return this.value + '%';
-                }
-
-            }
-        },
-        tooltip: {
-            valueSuffix: ' murders'
-        },
-        plotOptions: {
-            column: {
-                dataLabels: {
-                    enabled: true,
-                    formatter: function () {
-                        return this.y + '%';
-                    }
-                }
-            }
-        },
-				legend:{
-        enabled: false
-        },
-        credits: {
-            enabled: false
-        },
-        series: [{
-            name: 'Prior Arrests',
-            data: [9.1,9.4,25.5,21.6,12.6,7.6,14.2]
-
-        }]
-    }
-
-export {
-  hc003_table,
-  hc003_chart,
-  hc004_table,
-  hc004_chart,
-  hc005_chart,
-  hc_M_def_chart_150,
-  hc_M_def_chart_300,
-  hc_M_def_chart_400,
-  hc_M_def_chart_500,
-  hc_M_def_chart_600,
-  hc_M_def_chart_700,
-  hc_M_def_chart_800,
-  hc_M_def_chart_900,
-  hc_M_def_chart_1100
-};
